# frontend/generated/openapi-current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/investments": {
      "get": {
        "tags": [
          "investments"
        ],
        "summary": "Get Investments",
        "description": "Get all investments with calculated returns.",
        "operationId": "get_investments_api_investments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentsDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "investments"
        ],
        "summary": "Create Investment",
        "description": "Create a new investment.",
        "operationId": "create_investment_api_investments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Investment Api Investments Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investments/{investment_id}": {
      "put": {
        "tags": [
          "investments"
        ],
        "summary": "Update Investment",
        "description": "Update an investment.",
        "operationId": "update_investment_api_investments__investment_id__put",
        "parameters": [
          {
            "name": "investment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Investment Api Investments  Investment Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "investments"
        ],
        "summary": "Delete Investment",
        "description": "Delete an investment.",
        "operationId": "delete_investment_api_investments__investment_id__delete",
        "parameters": [
          {
            "name": "investment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Investment Api Investments  Investment Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/investments": {
      "get": {
        "tags": [
          "investments-workspace"
        ],
        "summary": "Get Investments",
        "description": "Get investments summary for the Investments Intelligence Workspace.\n\nReturns aggregated data matching InvestmentsViewModel format.",
        "operationId": "get_investments_api_v1_investments_get",
        "parameters": [
          {
            "name": "investment_types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Investment Types"
            }
          },
          {
            "name": "institutions",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Institutions"
            }
          },
          {
            "name": "statuses",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Statuses"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Investments Api V1 Investments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestmentCalculationStepDTO": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Step name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Step description"
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs",
            "description": "Input values"
          },
          "outputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Outputs",
            "description": "Output values"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description"
        ],
        "title": "InvestmentCalculationStepDTO",
        "description": "Calculation step in the investment derivation chain."
      },
      "InvestmentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "investment_type": {
            "type": "string",
            "title": "Investment Type"
          },
          "invested_paise": {
            "type": "integer",
            "title": "Invested Paise"
          },
          "current_value_paise": {
            "type": "integer",
            "title": "Current Value Paise"
          },
          "units": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Units"
          },
          "buy_price_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Price Paise"
          },
          "current_price_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price Paise"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "investment_type",
          "invested_paise",
          "current_value_paise"
        ],
        "title": "InvestmentCreate",
        "description": "Investment creation request."
      },
      "InvestmentEvidenceChainDTO": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Overall summary of the calculation"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/InvestmentEvidenceItemDTO"
            },
            "type": "array",
            "title": "Evidence",
            "description": "List of evidence items"
          },
          "calculation_steps": {
            "items": {
              "$ref": "#/components/schemas/InvestmentCalculationStepDTO"
            },
            "type": "array",
            "title": "Calculation Steps",
            "description": "Calculation chain steps"
          },
          "source_references": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source References",
            "description": "Source references for traceability"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score",
            "description": "Overall confidence (0-100)"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "confidence_score"
        ],
        "title": "InvestmentEvidenceChainDTO",
        "description": "Evidence chain for investment calculation."
      },
      "InvestmentEvidenceItemDTO": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Evidence type (holding, price, calculation)"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Human-readable summary"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Source reference"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "Confidence score (0-100)"
          }
        },
        "type": "object",
        "required": [
          "type",
          "summary",
          "source"
        ],
        "title": "InvestmentEvidenceItemDTO",
        "description": "Evidence item for investment calculation."
      },
      "InvestmentInsightDTO": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "positive",
              "warning",
              "info",
              "alert"
            ],
            "title": "Type",
            "description": "Insight type"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Severity",
            "description": "Insight severity"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable insight message"
          },
          "action_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Url",
            "description": "URL for detailed view or action"
          }
        },
        "type": "object",
        "required": [
          "type",
          "severity",
          "message"
        ],
        "title": "InvestmentInsightDTO",
        "description": "Insight about investment changes or patterns."
      },
      "InvestmentSummaryDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Investment identifier"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Investment name"
          },
          "type": {
            "type": "string",
            "enum": [
              "stocks",
              "mutual_funds",
              "bonds",
              "fd",
              "ppf",
              "gold",
              "other"
            ],
            "title": "Type",
            "description": "Investment type"
          },
          "institution": {
            "type": "string",
            "title": "Institution",
            "description": "Institution name"
          },
          "current_value_paise": {
            "type": "integer",
            "title": "Current Value Paise",
            "description": "Current value in paise"
          },
          "invested_paise": {
            "type": "integer",
            "title": "Invested Paise",
            "description": "Total invested in paise"
          },
          "returns_paise": {
            "type": "integer",
            "title": "Returns Paise",
            "description": "Absolute returns in paise"
          },
          "returns_percentage": {
            "type": "number",
            "title": "Returns Percentage",
            "description": "Returns percentage"
          },
          "returns_ytd_bps": {
            "type": "integer",
            "title": "Returns Ytd Bps",
            "description": "Year-to-date returns in basis points"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "closed",
              "matured"
            ],
            "title": "Status",
            "description": "Investment status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "institution",
          "current_value_paise",
          "invested_paise",
          "returns_paise",
          "returns_percentage",
          "returns_ytd_bps",
          "status"
        ],
        "title": "InvestmentSummaryDTO",
        "description": "Investment summary information."
      },
      "InvestmentUpdate": {
        "properties": {
          "units": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Units"
          },
          "current_price_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price Paise"
          },
          "current_value_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Value Paise"
          },
          "as_of_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "InvestmentUpdate",
        "description": "Investment update request."
      },
      "InvestmentsDTO": {
        "properties": {
          "investments": {
            "items": {
              "$ref": "#/components/schemas/InvestmentSummaryDTO"
            },
            "type": "array",
            "title": "Investments",
            "description": "List of investment summaries"
          },
          "total_value_paise": {
            "type": "integer",
            "title": "Total Value Paise",
            "description": "Total value across all investments in paise"
          },
          "total_invested_paise": {
            "type": "integer",
            "title": "Total Invested Paise",
            "description": "Total invested in paise"
          },
          "total_returns_paise": {
            "type": "integer",
            "title": "Total Returns Paise",
            "description": "Total returns in paise"
          },
          "investment_count": {
            "type": "integer",
            "title": "Investment Count",
            "description": "Total number of active investments"
          },
          "insights": {
            "items": {
              "$ref": "#/components/schemas/InvestmentInsightDTO"
            },
            "type": "array",
            "title": "Insights",
            "description": "List of insights about investments"
          },
          "evidence_chain": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestmentEvidenceChainDTO"
              },
              {
                "type": "null"
              }
            ],
            "description": "Evidence chain for explainability"
          }
        },
        "type": "object",
        "required": [
          "total_value_paise",
          "total_invested_paise",
          "total_returns_paise",
          "investment_count"
        ],
        "title": "InvestmentsDTO",
        "description": "Investments data transfer object.\n\nMonetary fields:\n- total_value_paise: Total value in paise (canonical)\n- total_invested_paise: Total invested in paise\n- total_returns_paise: Total returns in paise",
        "example": {
          "insights": [],
          "investment_count": 5,
          "investments": [],
          "total_invested_paise": 15000000,
          "total_returns_paise": 5000000,
          "total_value_paise": 20000000
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}